Convert containers to items and update page format
