+++ app/Models/Team.php
@@ -33,12 +33,6 @@ class Team extends Model
         return $this->belongsToMany(User::class)->withTimestamps();
     }
 
-    /** @return HasMany<Container, $this> */
-    public function containers(): HasMany
-    {
-        return $this->hasMany(Container::class);
-    }
-
     /** @return HasMany<TeamInvitation, $this> */
     public function invitations(): HasMany
     {
