+++ resources/views/layouts/app.blade.php
@@ -13,17 +13,9 @@
                         {{ __('Dashboard') }}
                     </flux:sidebar.item>
 
-                    @if (! request()->routeIs('inventory.*'))
                     <flux:sidebar.item icon="archive-box" :href="route('inventory.index')" :current="request()->routeIs('inventory.*')" wire:navigate>
                         {{ __('Inventory') }}
                     </flux:sidebar.item>
-                    @else
-                    <flux:sidebar.group expandable heading="Inventory" icon="archive-box" :expanded="request()->routeIs('inventory.*')">
-                        <flux:sidebar.item :href="route('inventory.index')" wire:navigate>{{ __('Overview') }}</flux:sidebar.item>
-                        <flux:sidebar.item :href="route('inventory.containers')" wire:navigate>{{ __('Containers') }}</flux:sidebar.item>
-                        <flux:sidebar.item :href="route('inventory.items')" wire:navigate>{{ __('Items') }}</flux:sidebar.item>
-                    </flux:sidebar.group>
-                    @endif
 
                     <flux:sidebar.item icon="book-open" :href="route('recipes.index')" :current="request()->routeIs('recipes.*')" wire:navigate>
                         {{ __('Recipes') }}
