--- resources/views/pages/inventory/heading.blade.php
@@ -1,5 +1,0 @@
-<div class="relative mb-6 w-full">
-    <flux:heading size="xl" level="1">{{ __('Inventory') }}</flux:heading>
-    <flux:subheading size="lg" class="mb-6">{{ __('Manage your posessions.') }}</flux:subheading>
-    <flux:separator variant="subtle" />
-</div>
